Silence many error messages

+++ include/AdamsIntegrator.hpp
@@ -179,7 +179,7 @@ void Adams_BM<T>::step(int n) {
     this->sys(tmp2, tmp, this->t[n+1]);
     tmp *= b_AM[0];
     // Now tmp goes back to being an aggregator
-    for (int i = 1; i < order; i++) {
+    for (size_t i = 1; i < order; i++) {
         this->sys(this->y[n-i+1], ydot, this->t[n-i+1]);
         ydot *= b_AM[i];
         tmp += ydot;
